Move config to the base class for tidyness
